+++ cypress/integration/2.10-fullRefund.spec.js
@@ -1,4 +1,8 @@
-import { requestRefund, singleProduct } from '../support/outputvalidation'
+import {
+  requestRefund,
+  singleProduct,
+  externalSeller,
+} from '../support/outputvalidation'
 import { loginViaCookies } from '../support/common/support.js'
 import { refund } from '../support/common/refund_apis.js'
 import { getRefundPayload } from '../support/refund_payload.js'
@@ -26,13 +30,16 @@ describe('Testing Cybersource transaction API for full refund', () => {
       total: requestRefund.getFullRefundTotal, // Amount
       title: 'full', // Refund Type for test case title
       env: requestRefund.fullRefundEnv, // variable name where we stored the orderid in node environment
+      externalSeller,
     },
-    getRefundPayload
+    getRefundPayload,
+    { sendInvoice: true, startHandling: true }
   )
 
   // verify cybersource transaction
   verifyRefundTid({
     prefix: 'fullRefund',
     paymentTransactionIdEnv,
+    payerAuth: true,
   })
 })